feat: moved boomkmarks reducer in redux folder
